requirements.txt
sentence-transformers
faiss-cpu
fastapi
numpy
pandas
pymilvus
uvicorn
torch
torchvision
Pillow
python-multipart